requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# AWS SDK
boto3==1.34.19
botocore==1.34.19

# SSH and Remote Execution
paramiko==3.4.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# WebSocket Support + Production WSGI Server
python-socketio==5.11.0
eventlet==0.35.1
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
validators==0.22.0

# Logging
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

#DB
sqlalchemy==2.0.25
alembic==1.13.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9   # PostgreSQL driver — needed for Docker Phase 1+
